move loader out
