+++ src/views/workflow-queries/workflow-queries.types.ts
@@ -1,4 +1,4 @@
-export type Props = {
+export type WorkflowQueryURLParams = {
   domain: string;
   cluster: string;
   workflowId: string;
